Fix invalid links 2 (#144)
* Update N1_SO2.md

* Update E12c_truck_detections_motorways.md

* Update E12d_truck_detections_primary_roads.md

* Update E13r_vessel_density_others.md

* Update E13e_cargo_boats_port.md

* Update E13f_fishing_boats_port.md

* Update E13m_maritime_traffic_passenger.md

* Update E13l_maritime_traffic_pleasure_craft.md

* Update E13i_maritime_traffic_search_rescue.md

* Update E13g_tanker_boats_port.md

* Update E13h_maritime_traffic_tug.md

* Update E13e_cargo_boats_port.md

* Update E13f_fishing_boats_port.md

* Add images and chart for maritime traffic analysis

Added images to illustrate the decrease in fishing and leisure boats in Genoa during the COVID-19 pandemic, and included a chart indicating lockdown measures.

* Add images of boat traffic impact during COVID-19

Added images showing the decrease in fishing and leisure boats in Genoa due to COVID-19 restrictions.

* Include images showing boat traffic changes during COVID-19

Added images to illustrate the impact of COVID-19 on maritime traffic.

* Update E13g_tanker_boats_port.md with COVID-19 impact visuals

Added images illustrating the decrease in fishing and leisure boats in Genoa and the increase in cargo boats in Gioia Tauro during COVID-19.

* Insert images showing boat traffic changes due to COVID-19

Added images to illustrate the impact of COVID-19 on fishing and leisure boats in Genoa, and cargo boats in Gioia Tauro.

* Add files via upload

* Update image URLs in E13e_cargo_boats_port.md

* Update E13e_cargo_boats_port.md

* Update E13e_cargo_boats_port.md

* Update IDEAS2_wildlife.md